+++ src/interfaces/selectCssClasses/selectCssClasses.interface.ts
@@ -1,0 +1,6 @@
+/**
+ * Css classes for Select
+ */
+export interface SelectCssClasses
+{
+}