+++ src/interfaces/selectOptions/selectOptions.interface.ts
@@ -4,37 +4,32 @@ import {SelectPluginTypes} from '../selectPluginTypes/selectPluginTypes.interfac
 /**
  * Describes select options used for Select
  */
-export interface SelectOptions<TValue = unknown> extends SelectBusOptions<TValue>
+export interface SelectOptions<TValue = unknown, TCssClasses = unknown> extends SelectBusOptions<TValue>
 {
     /**
-     * Indication whether Select should be initialized automaticaly during 'NgOnInit' phase
+     * Indication whether select should be initialized automaticaly during 'initialization' phase
      */
-    autoInitialize?: boolean;
+    autoInitialize: boolean;
 
     /**
-     * Css classes applied to ng select component
+     * Css classes applied to select component
      */
-    cssClasses?: unknown;
+    cssClasses: TCssClasses;
 
     /**
      * Object defining overrides for default plugins, default plugins can be also specified using DI
      */
-    plugins?: SelectPluginTypes;
+    plugins: SelectPluginTypes;
 
     /**
      * Indication whether is 'Popup' plugin displayd inside of Select (false) or directly in Body (true)
      */
-    absolute?: boolean;
+    absolute: boolean;
 
     /**
      * String that defines element in which should be absolute popup rendered, if not specified, body is used, used only if absolute is set to true, otherwise it has no effect
      *
      * Allows also css classes to be specified (div.body-box)
      */
-    containerElement?: string;
-
-    /**
-     * Indication whether force value check on initialization of select, this allows to change value of form control to value of ValueHandler
-     */
-    forceValueCheckOnInit?: boolean;
+    containerElement: string|undefined|null;
 }