+++ src/interfaces/visualPluginOptions/visualPluginOptions.interface.ts
@@ -8,5 +8,5 @@ export interface VisualPluginOptions<TCssClasses = unknown> extends PluginOption
     /**
      * Css classes applied to visual plugin (component), possible to override only part of classes
      */
-    cssClasses?: TCssClasses;
+    cssClasses: TCssClasses;
 }