+++ src/misc/types.ts
@@ -1,14 +1,14 @@
-import {Select, SelectOption} from '../interfaces';
+import {SelectApi, SelectOption} from '../interfaces';
 
 /**
  * Defintion of action that can be executed on Select
  */
-export type SelectAction<TValue = unknown> = (ngSelect: Select<TValue>) => void;
+export type SelectAction<TValue = unknown> = (ngSelect: SelectApi<TValue>) => void;
 
 /**
  * Definition of function that can be executed on Select and returns some data
  */
-export type SelectFunction<TResult = unknown, TValue = unknown> = (ngSelect: Select<TValue>) => TResult;
+export type SelectFunction<TResult = unknown, TValue = unknown> = (ngSelect: SelectApi<TValue>) => TResult;
 
 /**
  * Definition of function used for transformation of select option into display text